+++ src/main/java/net/tomasbot/mp/api/service/VideoFileScanner.java
@@ -8,8 +8,6 @@ import net.tomasbot.mp.model.Video;
 import net.tomasbot.mp.plugin.ffmpeg.metadata.FFmpegMetadata;
 import net.tomasbot.mp.plugin.ffmpeg.metadata.FFmpegStream;
 import org.apache.commons.io.FilenameUtils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Async;
@@ -18,13 +16,10 @@ import org.springframework.stereotype.Component;
 @Component
 public class VideoFileScanner implements FileMetadataScanner<Video> {
 
-  private static final Logger logger = LogManager.getLogger(VideoFileScanner.class);
-
   private final VideoService videoService;
   private final ThumbnailService thumbnailService;
-  private final TagService tagService;
+  private final PathTagService pathTagService;
   private final TranscodingService transcodingService;
-  private final InvalidFilesService invalidFilesService;
 
   @Value("${videos.storage-location}")
   private Path videoStorageLocation;
@@ -32,50 +27,44 @@ public class VideoFileScanner implements FileMetadataScanner<Video> {
   public VideoFileScanner(
       VideoService videoService,
       ThumbnailService thumbnailService,
-      TagService tagService,
-      TranscodingService transcodingService,
-      InvalidFilesService invalidFilesService) {
+      PathTagService pathTagService,
+      TranscodingService transcodingService) {
     this.videoService = videoService;
     this.thumbnailService = thumbnailService;
-    this.tagService = tagService;
+    this.pathTagService = pathTagService;
     this.transcodingService = transcodingService;
-    this.invalidFilesService = invalidFilesService;
   }
 
   @Override
   @Async("fileScanner")
-  public void scanFileMetadata(@NotNull Video video) {
+  public void scanFileMetadata(@NotNull Video video) throws IOException {
     final Path file = video.getFile();
-    try {
-      // metadata
-      final FFmpegMetadata metadata = transcodingService.getVideoMetadata(video);
-      List<FFmpegStream> streams = metadata.getStreams();
-      if (streams == null || streams.isEmpty()) {
-        String msg = String.format("Could not read video metadata for %s: no streams", file);
-        throw new IOException(msg);
-      }
-      video.setMetadata(metadata);
 
-      // title
-      final String title = transcodingService.getTitle(metadata);
-      if (title != null) video.setTitle(title);
-      else video.setTitle(FilenameUtils.getBaseName(file.toString()));
-
-      setVideoTags(video);
-
-      // thumbnails
-      videoService.save(video); // ensure video has ID
-      thumbnailService.generateVideoThumbnails(video);
-      videoService.save(video);
-    } catch (Throwable e) {
-      logger.error("Could not scan Video metadata: {}", e.getMessage(), e);
-      invalidFilesService.addInvalidFile(file, Video.class);
+    // metadata
+    final FFmpegMetadata metadata = transcodingService.getVideoMetadata(video);
+    List<FFmpegStream> streams = metadata.getStreams();
+    if (streams == null || streams.isEmpty()) {
+      String msg = String.format("Could not read video metadata for %s: no streams", file);
+      throw new IOException(msg);
     }
+    video.setMetadata(metadata);
+
+    // title
+    final String title = transcodingService.getTitle(metadata);
+    if (title != null) video.setTitle(title);
+    else video.setTitle(FilenameUtils.getBaseName(file.toString()));
+
+    setVideoTags(video);
+
+    // thumbnails
+    videoService.save(video); // ensure video has ID
+    thumbnailService.generateVideoThumbnails(video);
+    videoService.save(video);
   }
 
   private synchronized void setVideoTags(@NotNull Video video) {
     Path tagPath = videoStorageLocation.relativize(video.getFile());
-    final List<Tag> tags = tagService.getTags(tagPath);
+    final List<Tag> tags = pathTagService.getTagsFrom(tagPath);
     video.getTags().addAll(tags);
   }
 }