+++ src/main/java/net/tomasbot/mp/api/service/user/RemoveFavoriteException.java
@@ -1,0 +1,7 @@
+package net.tomasbot.mp.api.service.user;
+
+public class RemoveFavoriteException extends IllegalStateException {
+  public RemoveFavoriteException(Object o) {
+    super("Could not remove favorite on: " + o);
+  }
+}