+++ src/main/resources/static/js/mp.image.js
@@ -1,5 +1,4 @@
 import {
-    fetchFromRepoAt,
     getLinksArray,
     getLinkUrl,
     getViewportDimensions,